+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [10.2.0](https://github.com/webpack-contrib/less-loader/compare/v10.1.0...v10.2.0) (2021-10-18)
+
+
+### Features
+
+* use webpack logger for logging ([#444](https://github.com/webpack-contrib/less-loader/issues/444)) ([239c737](https://github.com/webpack-contrib/less-loader/commit/239c737e2ede1d17d83a6d11a6bd11211cf7d77d))
+
+## [10.1.0](https://github.com/webpack-contrib/less-loader/compare/v10.0.1...v10.1.0) (2021-10-11)
+
+
+### Features
+
+* add `link` field in schema ([#429](https://github.com/webpack-contrib/less-loader/issues/429)) ([8580731](https://github.com/webpack-contrib/less-loader/commit/858073159cc5aef320e7682798bc987f93ef4817))
+
+
+### Bug Fixes
+
+* only call `addDependency` on absolute paths ([fa11ce7](https://github.com/webpack-contrib/less-loader/commit/fa11ce7670ed8cae484e3435be05713deb199954))
+
+
 ### [10.0.1](https://github.com/webpack-contrib/less-loader/compare/v10.0.0...v10.0.1) (2021-07-02)